IK Model Constrain Improvement + Brush Target
Further constrained IK model to improve performance on 'Wrist 1'

Also added a Brush Target to improve path tracking performance when tracking more complex curve profiles and key-framing small scale action of the brush tip. It serves as a representation of the location of the brush/tool tip and is a useful offset alternative to changing the TCP offset location.

To use set Brush Target to follow your drawing path instead of IK target (which is now linked to Brush Target).

The current location of the brush target relative to the IK target is generic, but can be adjusted to match your tool tip location by moving the IK target in Object Mode relative to the Brush Target. Blender default dimensions are in meters.